requirements.txt
google-generativeai>=0.4.0
python-telegram-bot[job-queue]>=20.0
fastapi
uvicorn
python-dotenv
